+++ node/src/chain/acceptor.rs
@@ -309,7 +309,7 @@ impl<DB: database::DB, VM: vm::VMExecution, N: Network> Acceptor<N, DB, VM> {
         // hit the chain.
         let stake_calls =
             txs.iter().filter(|t| t.err.is_none()).filter_map(|t| {
-                match &t.inner.inner.payload().contract_call {
+                match &t.inner.inner.payload().contract_call() {
                     Some(call)
                         if (call.contract == STAKE_CONTRACT
                             && (call.fn_name == STAKE