+++ test-wallet/src/imp.rs
@@ -16,7 +16,7 @@ use alloc::vec::Vec;
 use dusk_bytes::Error as BytesError;
 use execution_core::{
     stake::{Stake, Unstake, Withdraw},
-    transfer::{ContractCall, Fee, Payload, Transaction},
+    transfer::{ContractCall, ContractExec, Fee, Payload, Transaction},
     BlsPublicKey as StakePublicKey, BlsScalar, JubJubScalar, Note,
     PhoenixError, PublicKey, SchnorrSecretKey, SecretKey, TxSkeleton, ViewKey,
     OUTPUT_NOTES,
@@ -343,7 +343,7 @@ where
     pub fn execute<Rng>(
         &self,
         rng: &mut Rng,
-        contract_call: ContractCall,
+        contract_exec: ContractExec,
         sender_index: u64,
         gas_limit: u64,
         gas_price: u64,
@@ -378,7 +378,7 @@ where
             outputs,
             fee,
             0,
-            Some(contract_call),
+            Some(contract_exec),
         )
         .map_err(Error::from_state_err)?;
 
@@ -505,7 +505,7 @@ where
             outputs,
             fee,
             value,
-            Some(contract_call),
+            Some(ContractExec::Call(contract_call)),
         )
         .map_err(Error::from_state_err)?;
 
@@ -589,7 +589,7 @@ where
             outputs,
             fee,
             deposit,
-            Some(call),
+            Some(ContractExec::Call(call)),
         )
         .map_err(Error::from_state_err)?;
 
@@ -674,7 +674,7 @@ where
             outputs,
             fee,
             deposit,
-            Some(call),
+            Some(ContractExec::Call(call)),
         )
         .map_err(Error::from_state_err)?;
 
@@ -740,7 +740,7 @@ fn new_unproven_tx<Rng: RngCore + CryptoRng, SC: StateClient>(
     outputs: [(Note, u64, JubJubScalar, [JubJubScalar; 2]); OUTPUT_NOTES],
     fee: Fee,
     deposit: u64,
-    call: Option<ContractCall>,
+    contract_exec: Option<ContractExec>,
 ) -> Result<UnprovenTransaction, SC::Error> {
     let nullifiers: Vec<BlsScalar> = inputs
         .iter()
@@ -766,7 +766,7 @@ fn new_unproven_tx<Rng: RngCore + CryptoRng, SC: StateClient>(
     let payload = Payload {
         tx_skeleton,
         fee,
-        contract_call: call,
+        contract_exec,
     };
     let payload_hash = payload.hash();